Split non-class APIs into individual files
