+++ src/lua_env/lua_api/lua_api_gui.hpp
@@ -1,0 +1,23 @@
+#pragma once
+#include "lua_api_common.hpp"
+
+namespace elona
+{
+namespace lua
+{
+
+namespace GUI
+{
+void txt(const std::string&);
+void txt_txtef(const std::string&, color_index_t);
+
+void txtnew();
+
+void txtef(color_index_t);
+
+
+void bind(sol::table&);
+}; // namespace GUI
+
+} // namespace lua
+} // namespace elona
