+++ src/lua_env/lua_api/lua_api_magic.hpp
@@ -1,0 +1,19 @@
+#pragma once
+#include "lua_api_common.hpp"
+
+namespace elona
+{
+namespace lua
+{
+
+namespace Magic
+{
+void cast_self(lua_character_handle, int, int, const position_t&);
+void cast(lua_character_handle, lua_character_handle, int, int);
+
+
+void bind(sol::table&);
+} // namespace Magic
+
+} // namespace lua
+} // namespace elona
