+++ src/lua_env/lua_api/lua_api_math.hpp
@@ -1,0 +1,18 @@
+#pragma once
+#include "lua_api_common.hpp"
+
+namespace elona
+{
+namespace lua
+{
+
+namespace Math
+{
+int clamp(int, int, int);
+
+
+void bind(sol::table&);
+} // namespace Math
+
+} // namespace lua
+} // namespace elona
