+++ src/lua_env/lua_api/lua_api_pos.hpp
@@ -1,0 +1,20 @@
+#pragma once
+#include "lua_api_common.hpp"
+
+namespace elona
+{
+
+namespace lua
+{
+
+namespace Pos
+{
+int dist(const position_t&, const position_t&);
+int dist_xy(int, int, int, int);
+
+
+void bind(sol::table&);
+} // namespace Pos
+
+} // namespace lua
+} // namespace elona
