+++ src/lua_env/lua_api/lua_api_rand.hpp
@@ -1,0 +1,22 @@
+#pragma once
+#include "lua_api_common.hpp"
+
+namespace elona
+{
+namespace lua
+{
+
+namespace Rand
+{
+int rnd(int n);
+
+bool one_in(int n);
+
+bool coinflip();
+
+
+void bind(sol::table&);
+}; // namespace Rand
+
+} // namespace lua
+} // namespace elona
