+++ src/lua_env/lua_api/lua_api_skill.hpp
@@ -1,0 +1,18 @@
+#pragma once
+#include "lua_api_common.hpp"
+
+namespace elona
+{
+namespace lua
+{
+
+namespace Skill
+{
+int level(int, lua_character_handle);
+
+
+void bind(sol::table&);
+} // namespace Skill
+
+} // namespace lua
+} // namespace elona
